requirements/dev.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/dev.in
blinker==1.7.0
    # via
    #   flask
    #   flask-debugtoolbar
click==8.1.7
    # via
    #   djlint
    #   flask
colorama==0.4.6
    # via djlint
cssbeautifier==1.15.1
    # via djlint
djlint==1.34.1
editorconfig==0.12.4
    # via
    #   cssbeautifier
    #   jsbeautifier
flask==3.0.2
    # via flask-debugtoolbar
flask-debugtoolbar==0.14.1
gunicorn==21.2.0
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
itsdangerous==2.1.2
    # via
    #   flask
    #   flask-debugtoolbar
jinja-partials==0.2.0
jinja2==3.1.3
    # via
    #   flask
    #   jinja-partials
jsbeautifier==1.15.1
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.24
    # via djlint
markupsafe==2.1.5
    # via
    #   flask-debugtoolbar
    #   jinja2
    #   werkzeug
mypy==1.9.0
mypy-extensions==1.0.0
    # via mypy
packaging==24.0
    # via
    #   flask-debugtoolbar
    #   gunicorn
pathspec==0.12.1
    # via djlint
python-dotenv==1.0.1
pyyaml==6.0.1
    # via djlint
regex==2023.12.25
    # via djlint
ruff==0.3.5
six==1.16.0
    # via
    #   cssbeautifier
    #   jsbeautifier
tqdm==4.66.2
    # via djlint
typing-extensions==4.10.0
    # via mypy
werkzeug==3.0.2
    # via
    #   flask
    #   flask-debugtoolbar
